pad: Expose Analog mode button.

+++ pcsx2/PAD/Host/KeyStatus.h
@@ -28,7 +28,7 @@ private:
 		u8 rx, ry;
 	};
 
-	u16 m_button[GAMEPAD_NUMBER];
+	u32 m_button[GAMEPAD_NUMBER];
 	u8 m_button_pressure[GAMEPAD_NUMBER][MAX_KEYS];
 	PADAnalog m_analog[GAMEPAD_NUMBER];
 	float m_axis_scale[GAMEPAD_NUMBER];
@@ -44,7 +44,7 @@ public:
 	__fi float GetVibrationScale(u32 pad, u32 motor) const { return m_vibration_scale[pad][motor]; }
 	__fi void SetVibrationScale(u32 pad, u32 motor, float scale) { m_vibration_scale[pad][motor] = scale; }
 
-	u16 GetButtons(u32 pad);
+	u32 GetButtons(u32 pad);
 	u8 GetPressure(u32 pad, u32 index);
 };
